Final Version
This code will output both total number of requests in log and total number of requests in 1995. It will also generate a local_copy.log. However you will need to change the file path to where the local_copy.log is created to be able to access the local disk copy. For more information please read the ReadMe.txt.